Check if update-manager release-upgrades configuration file is present before editing (#996)

+++ setup/system.sh
@@ -122,8 +122,10 @@ apt_install python3 python3-dev python3-pip \
 # ### Suppress Upgrade Prompts
 # Since Mail-in-a-Box might jump straight to 18.04 LTS, there's no need
 # to be reminded about 16.04 on every login.
-tools/editconf.py /etc/update-manager/release-upgrades Prompt=never
-rm -f /var/lib/ubuntu-release-upgrader/release-upgrade-available
+if [ -f /etc/update-manager/release-upgrades ]; then
+	tools/editconf.py /etc/update-manager/release-upgrades Prompt=never
+	rm -f /var/lib/ubuntu-release-upgrader/release-upgrade-available
+fi
 
 # ### Set the system timezone
 #
